+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Increase number of imported radars by fetch the next batch #6 Idea thanks to @futuretap
+
 ## [1.0.0(7)] - 2018-07-09
 ### Added
 - User agent in Open Radar API request #4